type of GfxElement->colortable was changed to UINT32 in MAME37b12, thus had to change stylus_color_table[] to UINT32 too

+++ mess/vidhrdw/avigo.c
@@ -32,7 +32,7 @@ static int stylus_y;
 
 
 /* colour table filled in from avigo colour table*/
-static unsigned short stylus_color_table[3] = {0,0,0};
+static UINT32 stylus_color_table[3] = {0,0,0};
 
 static struct GfxLayout pointerlayout =
 {
